move library/cpp/threading/atomic to library/cpp/deprecated/atomic_bool commit_hash:0981fb113212b130cb676672e125982188e4fd20
