+++ library/cpp/deprecated/atomic_bool/bool.cpp
