+++ library/cpp/deprecated/atomic_bool/bool.h
