+++ library/cpp/deprecated/atomic_bool/ut/ya.make
@@ -1,0 +1,7 @@
+UNITTEST_FOR(library/cpp/deprecated/atomic_bool)
+
+SRCS(
+    bool_ut.cpp
+)
+
+END()
