+++ library/cpp/deprecated/atomic_bool/ya.make
