+++ library/cpp/neh/utils.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <library/cpp/threading/atomic/bool.h>
+#include <library/cpp/deprecated/atomic_bool/bool.h>
 
 #include <util/network/address.h>
 #include <util/system/thread.h>
