--- library/cpp/threading/atomic/ut/ya.make
@@ -1,7 +1,0 @@
-UNITTEST_FOR(library/cpp/threading/atomic)
-
-SRCS(
-    bool_ut.cpp
-)
-
-END()
